Use panel title context menu for contextMenu actions
